Move utility functions into utils
